serving_service/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.116.1
grpcio==1.74.0
grpcio-tools==1.74.0
protobuf==6.32.1
pydantic==2.11.9
pydantic_core==2.33.2
requests==2.32.5
urllib3==2.5.0
uvicorn==0.35.0

